move VariantExtractionImpl
